Change max_bytes_for_level_multiplier to double

+++ db/c.cc
@@ -1512,7 +1512,7 @@ void rocksdb_options_set_max_bytes_for_level_base(
 }
 
 void rocksdb_options_set_max_bytes_for_level_multiplier(
-    rocksdb_options_t* opt, int n) {
+    rocksdb_options_t* opt, double n) {
   opt->rep.max_bytes_for_level_multiplier = n;
 }
