+++ java/src/main/java/org/rocksdb/MutableColumnFamilyOptionsInterface.java
@@ -474,7 +474,7 @@ public interface MutableColumnFamilyOptionsInterface {
    * @param maxBytesForLevelBase maximum bytes for level base.
    *
    * @return the reference to the current option.
-   * @see #setMaxBytesForLevelMultiplier(int)
+   * @see #setMaxBytesForLevelMultiplier(double)
    */
   MutableColumnFamilyOptionsInterface setMaxBytesForLevelBase(
       long maxBytesForLevelBase);
@@ -506,7 +506,7 @@ public interface MutableColumnFamilyOptionsInterface {
    * @see #setMaxBytesForLevelBase(long)
    */
   MutableColumnFamilyOptionsInterface setMaxBytesForLevelMultiplier(
-      int multiplier);
+      double multiplier);
 
   /**
    * The ratio between the total size of level-(L+1) files and the total
@@ -517,7 +517,7 @@ public interface MutableColumnFamilyOptionsInterface {
    *     the total size of level-L files for all L.
    * @see #maxBytesForLevelBase()
    */
-  int maxBytesForLevelMultiplier();
+  double maxBytesForLevelMultiplier();
 
   /**
    * Different max-size multipliers for different levels.
